+++ src/Internal/Metadata/ParadeDbAnnotationNames.cs
@@ -7,6 +7,7 @@ internal static class ParadeDbAnnotationNames
     public const string IndexFieldKinds = "ParadeDB:IndexFieldKinds";
     public const string IndexFieldTokenizers = "ParadeDB:IndexFieldTokenizers";
     public const string IndexFieldAliases = "ParadeDB:IndexFieldAliases";
+    public const string IndexFieldOpclasses = "ParadeDB:IndexFieldOpclasses";
     public const string IndexSearchTokenizer = "ParadeDB:IndexSearchTokenizer";
     public const string IndexKeyField = "ParadeDB:IndexKeyField";
     public const string IndexFields = "ParadeDB:IndexFields";